fix: preserve decimal max-price config values
Allow max-price config inputs to keep decimal precision and add regression coverage for trip pricing calculations so non-round values like 12.5 km/l don't get truncated.

+++ config/carpoolear.php
@@ -36,10 +36,10 @@ return [
     'min_version_ios' => env('MIN_VERSION_IOS') ?: null,
     'module_seat_price_enabled' => env('MODULE_SEAT_PRICE_ENABLED', false),
     'module_max_price_enabled' => env('MODULE_MAX_PRICE_ENABLED', false),
-    'module_max_price_fuel_price' => (int) env('MODULE_MAX_PRICE_FUEL_PRICE', 1500),
-    'module_max_price_price_variance_tolls' => (int) env('MODULE_MAX_PRICE_PRICE_VARIANCE_TOLLS', 10),
-    'module_max_price_price_variance_max_extra' => (int) env('MODULE_MAX_PRICE_PRICE_VARIANCE_MAX_EXTRA', 15),
-    'module_max_price_kilometer_by_liter' => (int) env('MODULE_MAX_PRICE_KILOMETER_BY_LITER', 10),
+    'module_max_price_fuel_price' => (float) env('MODULE_MAX_PRICE_FUEL_PRICE', 1500),
+    'module_max_price_price_variance_tolls' => (float) env('MODULE_MAX_PRICE_PRICE_VARIANCE_TOLLS', 10),
+    'module_max_price_price_variance_max_extra' => (float) env('MODULE_MAX_PRICE_PRICE_VARIANCE_MAX_EXTRA', 15),
+    'module_max_price_kilometer_by_liter' => (float) env('MODULE_MAX_PRICE_KILOMETER_BY_LITER', 10),
 
     'manual_identity_validation_cost_cents' => (int) env('MANUAL_IDENTITY_VALIDATION_COST_CENTS', 0),
